+++ obadiah/src/main/java/net/zionsoft/obadiah/notification/PushNotificationReceiver.java
@@ -27,7 +27,7 @@ import android.support.v4.app.NotificationCompat;
 import android.support.v4.content.ContextCompat;
 import android.text.TextUtils;
 
-import com.crashlytics.android.Crashlytics;
+import com.google.firebase.crash.FirebaseCrash;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 import com.squareup.moshi.Moshi;
@@ -123,7 +123,7 @@ public class PushNotificationReceiver extends FirebaseMessagingService {
                     .setContentText(verse.text.text)
                     .setStyle(new NotificationCompat.BigTextStyle().bigText(verse.text.text));
         } catch (Exception e) {
-            Crashlytics.getInstance().core.logException(e);
+            FirebaseCrash.report(e);
             return false;
         }
         return true;